backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==21.1
djoser==2.1.0
gunicorn==23.0.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0.1
setuptools==65.0.0
python-dotenv==1.2.1